+++ meta/store.go
@@ -36,9 +36,8 @@ const ExecMagic = "EXEC"
 
 // Retention policy auto-create settings.
 const (
-	AutoCreateRetentionPolicyName     = "default"
-	AutoCreateRetentionPolicyReplicaN = 1
-	AutoCreateRetentionPolicyPeriod   = 0
+	AutoCreateRetentionPolicyName   = "default"
+	AutoCreateRetentionPolicyPeriod = 0
 )
 
 // Raft configuration.
@@ -186,7 +185,6 @@ func (s *Store) Open() error {
 	if s.id == 0 {
 		go s.init()
 	} else {
-		s.waitForLeader(10 * time.Second)
 		close(s.ready)
 	}
 
@@ -331,7 +329,7 @@ func (s *Store) init() {
 // Writes the id of the node to file on success.
 func (s *Store) createLocalNode() error {
 	// Wait for leader.
-	if err := s.waitForLeader(5 * time.Second); err != nil {
+	if err := s.WaitForLeader(5 * time.Second); err != nil {
 		return fmt.Errorf("wait for leader: %s", err)
 	}
 
@@ -354,8 +352,12 @@ func (s *Store) createLocalNode() error {
 	return nil
 }
 
-// waitForLeader sleeps until a leader is found or a timeout occurs.
-func (s *Store) waitForLeader(timeout time.Duration) error {
+// WaitForLeader sleeps until a leader is found or a timeout occurs.
+func (s *Store) WaitForLeader(timeout time.Duration) error {
+	if s.raft.Leader() != "" {
+		return nil
+	}
+
 	// Begin timeout timer.
 	timer := time.NewTimer(timeout)
 	defer timer.Stop()
@@ -614,13 +616,25 @@ func (s *Store) CreateDatabase(name string) (*DatabaseInfo, error) {
 	}
 
 	if s.retentionAutoCreate {
+		// Read node count.
+		// Retention policies must be fully replicated.
+		var nodeN int
+		if err := s.read(func(data *Data) error {
+			nodeN = len(data.Nodes)
+			return nil
+		}); err != nil {
+			return nil, fmt.Errorf("read: %s", err)
+		}
+
+		// Create a retention policy.
 		rpi := NewRetentionPolicyInfo(AutoCreateRetentionPolicyName)
-		rpi.ReplicaN = AutoCreateRetentionPolicyReplicaN
+		rpi.ReplicaN = nodeN
 		rpi.Duration = AutoCreateRetentionPolicyPeriod
 		if _, err := s.CreateRetentionPolicy(name, rpi); err != nil {
 			return nil, err
 		}
 
+		// Set it as the default retention policy.
 		if err := s.SetDefaultRetentionPolicy(name, AutoCreateRetentionPolicyName); err != nil {
 			return nil, err
 		}